requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
httpx==0.27.0
chromadb==0.5.0
sentence-transformers==2.7.0
pypdf==4.2.0
google-generativeai==0.5.4
python-dotenv==1.0.1